infrastructure/testharness.js/single-page-test-silent-before-results.html

<!DOCTYPE HTML>
<html>
<head>
<title>Harness does not modify DOM prior to completion</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body><script>
'use strict';
/**
 * This property is useful for tests which concern events related to the
 * initial rendering of the document, e.g. paint-timing.
 */
setup({single_test: true});
addEventListener('load', function() {
  step_timeout(function() {
    var childNodes = Array.prototype.slice.call(document.body.childNodes);
    assert_equals(childNodes.length, 2);
    assert_equals(childNodes[0].tagName, 'SCRIPT');
    assert_equals(childNodes[1].nodeType, 3);
    assert_equals(childNodes[1].data, '\n');
    done();
  }, 300);
});
</script></body></html>
